happy e2e script
