+++ e2e/btc-cli.ts
